+++ src/app/api/users/[id]/route.ts
@@ -2,10 +2,17 @@ import { NextResponse } from "next/server";
 import { PrismaClient } from "@prisma/client";
 import type { User } from "@prisma/client";
 import bcrypt from 'bcrypt'
+import { getServerSession } from "next-auth"; // Anda perlu menginstal next-auth untuk ini
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 
 const prisma = new PrismaClient();
 
 export const DELETE = async (req: Request, { params }: { params: { id: string }}) => {
+    // Dapatkan sesi pengguna
+    const session = await getServerSession(authOptions);
+    if (session?.user.role !== "Admin") {
+        return NextResponse.json({ error: "Method Not Allowed" }, { status: 401 });
+    }
     const user = await prisma.user.delete({
          where: {
              id: Number(params.id)
@@ -16,9 +23,14 @@ export const DELETE = async (req: Request, { params }: { params: { id: string }}
 
 export const PATCH = async (req: Request, { params }: { params: { id: string }}) => {
     const body: User = await req.json();
-const hashedPassword = await bcrypt.hash(body.password, 10);
-const expireDate = new Date(body.expireDate);
-const data: Partial<User> = {
+    const hashedPassword = await bcrypt.hash(body.password, 10);
+    const expireDate = new Date(body.expireDate);
+    // Dapatkan sesi pengguna
+    const session = await getServerSession(authOptions);
+    if (session?.user.role !== "Admin") {
+        return NextResponse.json({ error: "Method Not Allowed" }, { status: 401 });
+    }
+    const data: Partial<User> = {
     username: body.username,
     status: body.status,
     role: body.role,
